requirements.txt
asgiref
dataclasses
Django~=4.2.2
django-ninja
pydantic~=1.10.9
pytz
sqlparse
typing-extensions

django-log-request-id
colorlog==4.0.2
PyMySQL~=1.0.3
django-redis
